requirements.txt
flask==3.0.2
flask-cors==4.0.0
pyngrok==7.0.5
google-generativeai==0.3.2 